momclient/requirements.txt
grpcio-tools==1.51.3
grpcio==1.51.3
